Fix nav links scrolling too far down
Previously this issue was fixed by manually adjusting the
scroll-margin-top property of the various sections based on whether or
not the navbar was open on mobile. This was due to the navbar being set
to sticky which pushed the pages content down to make room for it. Now,
the navbar is set to fixed and the scroll-margin-top can be statically
set to the height of the navbar to ensure proper position when scrolling
to a section via a link.